refactor: Move java files to proper folder according to package name
Move the java files to a new folder to reflect the change in the package name. The package is now org.hyde
