+++ src/main/java/org/hyde/Build.java
@@ -1,21 +1,22 @@
+package org.hyde;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 import java.util.concurrent.Callable;
 
-@Command(name = "clean")
+@Command(name = "build")
 
-class Clean implements Callable<Integer> {
+class Build implements Callable<Integer> {
 
    @Override
    public Integer call() {
-      System.out.println("Commande 'clean'");
+      System.out.println("Commande 'build'");
       return 0;
    }
 
    public static void main(String... args) {
-      int exitCode = new CommandLine(new Clean()).execute(args);
+      int exitCode = new CommandLine(new Build()).execute(args);
       System.exit(exitCode);
    }
 }
