+++ src/main/java/org/hyde/Clean.java
@@ -1,21 +1,22 @@
+package org.hyde;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 import java.util.concurrent.Callable;
 
-@Command(name = "new")
+@Command(name = "clean")
 
-class New implements Callable<Integer> {
+class Clean implements Callable<Integer> {
 
    @Override
    public Integer call() {
-      System.out.println("Commande 'new'");
+      System.out.println("Commande 'clean'");
       return 0;
    }
 
    public static void main(String... args) {
-      int exitCode = new CommandLine(new New()).execute(args);
+      int exitCode = new CommandLine(new Clean()).execute(args);
       System.exit(exitCode);
    }
 }
