+++ src/main/java/org/hyde/Main.java
@@ -1,6 +1,7 @@
+package org.hyde;
 
-class Main {
+public class Main {
    public static void main(String... args) {
-
+      
    }
 }
