+++ src/main/java/org/hyde/New.java
@@ -1,21 +1,22 @@
+package org.hyde;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 import java.util.concurrent.Callable;
 
-@Command(name = "build")
+@Command(name = "new")
 
-class Build implements Callable<Integer> {
+class New implements Callable<Integer> {
 
    @Override
    public Integer call() {
-      System.out.println("Commande 'build'");
+      System.out.println("Commande 'new'");
       return 0;
    }
 
    public static void main(String... args) {
-      int exitCode = new CommandLine(new Build()).execute(args);
+      int exitCode = new CommandLine(new New()).execute(args);
       System.exit(exitCode);
    }
 }
